+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.5.0
  * Added ability to use custom date format for tooltip dates
+ * Added ability to use extra information in the tooltip
 ## 1.4.0
  * Added new date types for x-axis(Quarter, Hours, Minutes, Seconds)
 ## 1.3.0